Re-add scan functionality
