+++ src/main.py
@@ -1,151 +1,473 @@
-#import base64
+import base64
 import vlc
-import time
 import glob
 from natsort import natsorted
 import curses
-#import RPi.GPIO as GPIO
-#import pn532.pn532 as nfc
-#from pn532 import *
+import RPi.GPIO as GPIO
+import pn532.pn532 as nfc
+from pn532 import *
+from pathlib import Path
+import time
 
-"""
-# UART connection
-pn532 = PN532_UART(debug=False, reset=20)
 
-# Configure PN532 to communicate with MiFare cards
-pn532.SAM_configuration()
+class Context:
+    def __init__(self, initial_state, player, nfc_reader):
+        self.state = None
+        
+        self.player = player
+        self.nfc_reader = nfc_reader
 
-# Scan until card is detected
-print("Listening! Please present card")
-while True:
-    # Check if a card is available to read
-    uid = pn532.read_passive_target(timeout=0.5)
-    print('.', end="")
+        self.set_state(initial_state)
+
+    # -------------------------
+    # State Management
+    # -------------------------
+
+    def set_state(self, new_state):
+        if self.state is not None:
+            self.state.exit(self)
+            
+        self.state = new_state
+        self.state.enter(self)
+
+    # -------------------------
+    # Input Management
+    # -------------------------
+
+    def handle_key(self, key):
+        self.state.handle_key(self, key)
+
+    def hanfle_nfc(self, nfc_reader):
+        self.state.handle_nfc(nfc_reader)
+
+    # -------------------------
+    # Player Helpers
+    # -------------------------
+
+    def load_album(self, album_name):
+        self.player.load_album(album_name)
     
-    # Try again if no card is available.
-    if uid is not None:
-        break
+    def unload_album(self):
+        self.player.unload_album()
 
-print('Found card with UID:', [hex(i) for i in uid])
+    def play(self):
+        self.player.play()
 
-# Set key
-key_a = b'\xFF\xFF\xFF\xFF\xFF\xFF'
-data_blocks = []
+    def pause(self):
+        self.player.pause()
 
-for x in range (2):
-    # We need blocks one and two so increment index
-    index = x + 1
+    def stop(self):
+        self.player.stop()
+
+    def next_track(self):
+        self.player.next_track()
+
+    def previous_track(self):
+        self.player.previous_track()
+
+    def has_song_ended(self):
+        return self.player.has_song_ended()
     
-    # Authenticate and read block
-    pn532.mifare_classic_authenticate_block(uid, block_number=index, key_number=nfc.MIFARE_CMD_AUTH_A, key=key_a)
-    block = pn532.mifare_classic_read_block(index)
+    # -------------------------
+    # NFC Helpers
+    # -------------------------
+
+    def scan(self):
+        return self.nfc_reader.scan()
+
+    # -------------------------
+    # Getters
+    # -------------------------
+
+    def get_state_name(self):
+        return self.state.get_state_name()
     
-    # Filter block and append to list
-    block =  block.replace(b'\x00', b'').decode('utf-8')
-    data_blocks.append(block)
+    def get_playlist(self):
+        return self.player.get_playlist()
+    
+    def get_current_track(self):
+        return self.player.get_current_track()
 
-# Append blocks 1 and 2 
-album_name = data_blocks[0] + data_blocks[1] # Assume we always read two blocks
-"""
 
-# Delete before commit
-album_name = "Test"
+class State:
+    def enter(self, context):
+        """
+        We call this when the state becomes active
+        Use this when we want to start scanning, playing etc...
+        """
+        pass
 
-# Get all songs in album - Order them into queue
-mp3_files = list(
-    glob.iglob('../albums/'+ album_name +'/*.mp3')
-)
-playlist = natsorted(mp3_files)
+    def exit(self, context):
+        """
+        We call this when we leave the state
+        Use this for cleanup when needed
+        """
+        pass
 
-current_track = 0
+    def handle_key(self, context, key):
+        """
+        Handle keyboard input
+        """
+        pass
 
-player = vlc.MediaPlayer(playlist[current_track])
-player.play()
+    def handle_nfc(self, context, nfc_reader):
+        """
+        Handle NFC events
+        """
+        pass
 
-paused = False
+    def get_state_name(self):
+        """
+        Debug function
+        """
+        pass
 
-# Set up input
-screen = curses.initscr()
-curses.cbreak()
-screen.keypad(True)
 
-try:
-    while True:
+class Playing(State):
+    def enter(self, context):
+        context.play()
+    
+    def exit(self, context):
+        """
+        Ignore
+        """
+        pass
+
+    def handle_key(self, context, key):
+        if key == "ENTER":
+            context.set_state(Paused())
+        elif key == "LEFT":
+            context.previous_track()
+            context.play()
+        elif key == "RIGHT":
+            has_next = context.get_current_track() < len(context.get_playlist()) - 1
+            if has_next: 
+                context.next_track()
+                context.play()
+            else:
+                context.set_state(Stopped())
+        elif key == "UP":
+            context.set_state(Scanning())
+        elif key == "DOWN":
+            context.set_state(Stopped())
+
+    def get_state_name(self):
+        return "Playing"
+
+
+class Paused(State):
+    def enter(self, context):
+        context.pause()
+
+    def exit(self, contex):
+        """
+        Ignore
+        """
+        pass
+
+    def handle_key(self, context, key):
+        if key == "ENTER":
+            context.set_state(Playing())
+        elif key == "LEFT":
+            context.previous_track()
+        elif key == "RIGHT":
+            has_next = context.get_current_track() < len(context.get_playlist()) - 1
+            if has_next: 
+                context.next_track()
+            else:
+                context.set_state(Stopped())
+        elif key == "UP":
+            context.set_state(Scanning())
+        elif key == "DOWN":
+            context.set_state(Stopped())
+
+    def get_state_name(self):
+        return "Paused"
+
+
+class Scanning(State):
+    def enter(self, context):
+        context.stop()
+        album_name = context.scan()
+        context.load_album(album_name)
+        #context.load_album("VoulezVous")
+        context.set_state(Playing())
+
+    def exit(self, context):
+        pass
+
+    def handle_key(self, context, key):
+        pass
+
+    def get_state_name(self):
+        return "Scanning"
+
+
+class Stopped(State):
+    def enter(self, context):
+        context.stop()
+        context.unload_album()
+
+    def exit(self, context):
+        pass
+
+    def handle_key(self, context, key):
+        if key == "UP":
+            context.set_state(Scanning())
+
+    def get_state_name(self):
+        return "Stopped"
+
+
+class Player:
+    def __init__(self):
+        self.playlist = []
+        self.current_track = 0
+        self.player = None
+
+    def load_album(self, album_name):
+        # Get all songs in album - Order them into queue
+        mp3_files = list(
+            glob.iglob('../albums/'+ album_name +'/*.mp3')
+        )
+
+        self.playlist = natsorted(mp3_files)
+        self.current_track = 0
+        self.player = vlc.MediaPlayer(
+            self.playlist[self.current_track]
+        )
+        
+    def unload_album(self):
+        if self.playlist != []:
+            self.playlist = []
+
+    def play(self):
+        if self.player:
+            self.player.play()
+
+    def pause(self):
+        if self.player:
+            self.player.pause()
+
+    def stop(self):
+        if self.player:
+            self.player.stop()
+
+    def next_track(self):
+        self.stop()
+
+        self.current_track += 1
+
+        self.player = vlc.MediaPlayer(
+            self.playlist[self.current_track]
+        )
+
+    def previous_track(self):
+        # If over 5 seconds has elapsed restart the song otherwise go to previous track
+        if self.player and self.player.get_time() > 5000:
+            self.player.set_time(0)
+            return
+
+        # To Previous Track if we are on first track then restart first track
+        self.stop()
+
+        self.current_track -= 1
+
+        if self.current_track < 0:
+            self.current_track = 0
+            
+        self.player = vlc.MediaPlayer(
+            self.playlist[self.current_track]
+        )
+
+    def has_song_ended(self):
+        return ( self.player and self.player.get_state() == vlc.State.Ended )
+
+    def get_playlist(self):
+        return self.playlist
+
+    def get_current_track(self):
+        return self.current_track
+
+
+class NFCReader:
+    def __init__(self):
+        # UART connection
+        self.pn532 = PN532_UART(debug=False, reset=20)
+
+        # Configure PN532 to communicate with MiFare cards
+        self.pn532.SAM_configuration()
+    
+    def scan(self):
+        while True:
+            # Check if a card is available to read
+            uid = self.pn532.read_passive_target(timeout=0.5)
+    
+            # Try again if no card is available.
+            if uid is not None:
+                break
+            
+        # Set key
+        key_a = b'\xFF\xFF\xFF\xFF\xFF\xFF'
+        data_blocks = []
+
+        for x in range (2):
+            # We need blocks one and two so increment index
+            index = x + 1
+    
+            # Authenticate and read block
+            self.pn532.mifare_classic_authenticate_block(uid, block_number=index, key_number=nfc.MIFARE_CMD_AUTH_A, key=key_a)
+            block = self.pn532.mifare_classic_read_block(index)
+    
+            # Filter block and append to list
+            block =  block.replace(b'\x00', b'').decode('utf-8')
+            data_blocks.append(block)
+
+        # Append blocks 1 and 2 
+        album_name = data_blocks[0] + data_blocks[1] # Assume we always read two blocks
+        return album_name
+
+
+def format_playlist(current_track, playlist):
+    # Build output string from extracted file names in playlist - Highlight current track
+    playlist_output = ""
+    for index, song in enumerate(playlist):
+        song_path = Path(song)
+        formatted_song = song_path.stem
+
+        if (index == current_track):
+            formatted_song = "> " + formatted_song + " <"
+
+        playlist_output += formatted_song + "\n"
+
+    return playlist_output
+
+
+def build_output(app, confirmation_action):
+    # Return confirmation message if confirmation action is active
+    if confirmation_action is not None:
+        return "Are you sure you want to " + confirmation_action + " [Y / N]"
+    
+    # If user is not performing a confirmation action build playback UI
+    state = app.get_state_name()
+    current_track = app.get_current_track()
+    playlist = app.get_playlist()
+
+    playlist_output = format_playlist(current_track, playlist) if playlist != [] else "No album loaded. Press the up arrow to scan for an album"
+    output = state + "\n" + playlist_output
+
+    return output
+
+
+def print_output(screen, output):
+    # Clear screen, set output variables and transform output string into array
+    screen.clear()
+    max_y, max_x = screen.getmaxyx()
+    lines = (output).split("\n")
+    
+    # Abort output if if output exceeds terminal size to prevent crashing.
+    for i, line in enumerate(lines):
+        if i >= max_y -1:
+            break
+        screen.addstr(i, 0, line[:max_x -1])
+
+    screen.refresh()
+
+
+def handle_output(app, screen, confirmation_action, previous_output):
+    # Build output based on app state and whether a confirmation action has been chosen
+    output = build_output(app, confirmation_action)
+
+    # Do not unecessarily re-render - only render if output has been changed since last render
+    if output != previous_output:
+        print_output(screen, output)
+
+    return output
+
+
+def handle_action_input(app, key):
+    confirmation_action = None
+
+    if key == 10:
+        app.handle_key("ENTER")
+    elif key == curses.KEY_LEFT:
+        app.handle_key("LEFT")
+    elif key == curses.KEY_RIGHT:
+        app.handle_key("RIGHT")
+    elif key == curses.KEY_UP:
+        app.handle_key("UP")
+    elif key == curses.KEY_DOWN:
+        app.handle_key("DOWN")
+    elif key == ord("q") or key == ord("Q"):
+        confirmation_action = "QUIT"
+    
+    return confirmation_action
+
+
+def handle_confirmation_input(key, confirmation_action):
+    # If is "yes" continue with now confirmed action
+    if key == ord("y") or key == ord("Y"):
+        if confirmation_action == "QUIT":
+            confirmation_action = "QUIT_CONFIRMED"
+        elif confirmation_action == "SHUTDOWN":
+            pass    # Shutdown - Do nothing right now, I do not want to test this yet
+
+    # Otherwise reset confirmation_action to none
+    elif key == ord("n") or key == ord("N"):
+        confirmation_action = None
+    else:
+        pass    # Print bad input
+
+    return confirmation_action
+
+
+def main(screen):
+    # Initialise application
+    curses.cbreak()
+    screen.keypad(True)
+    curses.noecho()
+
+    # Enable non-blocking input
+    screen.nodelay(True)
+
+    # Initialise classes
+    player = Player()
+    nfc_reader = NFCReader()
+    app = Context(Stopped(), player, nfc_reader)
+
+    # Used to handle special cases in app such as Quit and Shut Down - Used by handle_output() to render action confirmation screen 
+    confirmation_action = None
+
+    # Store previous output to conditionally render UI if change is made - This stops UI flicker
+    previous_output = None
+
+    # Main loop
+    running = True
+    while running:
         key = screen.getch()
 
-        if key == 10: # Enter
-            if paused:
-                player.pause()
-                paused = False
-
+        # If input detected handle it
+        if key != -1:
+            if confirmation_action is None:
+                confirmation_action = handle_action_input(app, key)
             else:
-                player.pause()
-                paused = True
+                confirmation_action = handle_confirmation_input(key, confirmation_action)
+                
+                if confirmation_action == "QUIT_CONFIRMED":
+                    running = False
 
-        elif key == curses.KEY_LEFT:   
-            # If over 5 seconds has elapsed restart the song otherwise go to previous track
-            if player.get_time() > 5000:
-                player.set_time(0)
-                continue
+        # If song has ended play the next track unless there is no next track in which case enter Stopped State
+        if app.has_song_ended():
+            has_next = app.get_current_track() < len(app.get_playlist()) - 1
 
-            # To Previous Track
-            player.stop()
+            if has_next: 
+                app.next_track()
+                app.play()
+            else:
+                app.set_state(Stopped())
 
-            current_track -= 1
-
-            if current_track < 0:
-                current_track = 0
-            
-            player = vlc.MediaPlayer(
-                playlist[current_track]
-            )
-
-            player.play()
-
-        elif key == curses.KEY_RIGHT:
-            player.stop()
-
-            current_track += 1
-
-            if current_track >= len(playlist):
-                screen.addstr("Album finished!")
-                break
-
-            player = vlc.MediaPlayer(
-                playlist[current_track]
-            )
-
-            player.play()
-
-        elif key == curses.KEY_DOWN:
-            break
-
-finally:
-    curses.nocbreak()
-    screen.keypad(False)
-    curses.echo()
-    curses.endwin()
-
-print("Playback ended!")
+        # Build output based on state of app / confirmation_action and print
+        previous_output = handle_output(app, screen, confirmation_action, previous_output)
 
 
-# handle_scan()
-
-# handle_pause()
-
-# handle_back()
-
-# handle_forward()
-
-# handle_shutdown()
-
-# handle_quit()
-
-# update_state(state)
-
-# album = handle_scan()
-# album.play()
-
-# while true:
-    # Read controls and perform desired action
-    # Render UI (clear everything and output global variables)
+curses.wrapper(main)
